// packages/investing/src/predictos/ai/types.ts
export interface OpenAIOutputText {
  type: "output_text";
  text: string;
  annotations?: unknown[];
}

export interface OpenAIMessage {
  content: OpenAIOutputText[];
  id: string;
  role: "assistant";
  type: "message";
  status: "completed" | "failed" | "pending";
}
// ...
export interface GrokAnnotation {
  type: "url_citation";
  url: string;
}

export interface GrokOutputText {
  type: "output_text";
  text: string;
  logprobs?: unknown[];
  annotations?: GrokAnnotation[];
}

export interface GrokCustomToolCall {
  call_id: string;
  input: string;
  name: string;
  type: "custom_tool_call";
  id: string;
  status: "completed" | "failed" | "pending";
}

export interface GrokMessage {
  content: GrokOutputText[];
  id: string;
  role: "assistant";
  type: "message";
  status: "completed" | "failed" | "pending";
}